In pom.xml version number for maven surefire moved to properties and  replaced by a variable; added folders for data provider and test classes; added general class BaseTest for tests; fixed class TestDivideByZeroLongFunction
