+++ pom.xml
@@ -18,6 +18,7 @@
         <testng.version>7.4.0</testng.version>
         <assertj-core.version>3.4.1</assertj-core.version>
         <calculator.version>1.0.0</calculator.version>
+        <maven-surefire-plugin.version>3.0.0-M5</maven-surefire-plugin.version>
     </properties>
 
     <build>
@@ -25,7 +26,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>3.0.0-M5</version>
+                <version>${maven-surefire-plugin.version}</version>
                 <configuration>
                     <suiteXmlFiles>
                         <suiteXmlFile>${maven-surefire-plugin.suiteXmlFile}</suiteXmlFile>
