+++ src/test/resources/testing/hw1.runAddAndSubtractTest.xml
@@ -8,8 +8,8 @@
     </groups>
 
     <classes>
-      <class name="com.epam.tc.hw1.TestAddDoubleFunction"/>
-      <class name="com.epam.tc.hw1.TestSubtractDoubleFunction"/>
+      <class name="com.epam.tc.hw1.tests.TestAddDoubleFunction"/>
+      <class name="com.epam.tc.hw1.tests.TestSubtractDoubleFunction"/>
     </classes>
   </test>
 </suite>
