+++ src/test/resources/testing/hw1.runMultiplyAndDivideTest.xml
@@ -8,9 +8,9 @@
     </groups>
 
     <classes>
-      <class name="com.epam.tc.hw1.TestDivideDoubleFunction"/>
-      <class name="com.epam.tc.hw1.TestForDivideByZeroLongFunction"/>
-      <class name="com.epam.tc.hw1.TestMultiplyDoubleFunction"/>
+      <class name="com.epam.tc.hw1.tests.TestDivideDoubleFunction"/>
+      <class name="com.epam.tc.hw1.tests.TestDivideByZeroLongFunction"/>
+      <class name="com.epam.tc.hw1.tests.TestMultiplyDoubleFunction"/>
     </classes>
   </test>
 </suite>
